+++ server/mock/store.go
@@ -159,6 +159,10 @@ func (ms multiStore) RollbackToVersion(version int64) error {
 	panic("not implemented")
 }
 
+func (ms multiStore) LatestVersion() int64 {
+	panic("not implemented")
+}
+
 var _ sdk.KVStore = kvStore{}
 
 type kvStore struct {